Added initial support to maintain and output metrics.

+++ src/renderer.cc
@@ -796,7 +796,7 @@ bool FileRenderer::OutputJDirectory(ParsedDirectory* dir) {
 
         WriteJsonKeyValue(&writer, "href", descriptor.HtmlPath());
         WriteJsonKeyValue(&writer, "mtime", ctime(&descriptor.mtime));
-        WriteJsonKeyValue(&writer, "size", descriptor.size);
+        WriteJsonKeyValue(&writer, "size", static_cast<uint64_t>(descriptor.size));
       }
     }
